Move some render functions to utils
